Fixing deleting in-progress task

+++ repository/src/main/java/org/apache/atlas/repository/store/graph/v2/EntityGraphMapper.java
@@ -2887,6 +2887,9 @@ public class EntityGraphMapper {
                                         && t.getType().equals(CLASSIFICATION_PROPAGATION_ADD))
                                 .collect(Collectors.toList());
                         for (AtlasTask entityClassificationPendingTask: entityClassificationPendingTasks) {
+                            if (AtlasTask.Status.IN_PROGRESS.name().equals(entityClassificationPendingTask.getStatus())) {
+                                throw new AtlasBaseException("Classification propagation is going on, Try attaching after some time");
+                            }
                             String taskGuid = entityClassificationPendingTask.getGuid();
                             taskManagement.deleteByGuid(taskGuid, TaskManagement.DeleteType.SOFT);
                             AtlasGraphUtilsV2.deleteProperty(entityVertex, PENDING_TASKS_PROPERTY_KEY, taskGuid);
